refactor: move events to separate package
