+++ android/src/main/java/com/swmansion/rnscreens/ScreenViewManager.java
@@ -6,6 +6,13 @@ import com.facebook.react.module.annotations.ReactModule;
 import com.facebook.react.uimanager.ThemedReactContext;
 import com.facebook.react.uimanager.ViewGroupManager;
 import com.facebook.react.uimanager.annotations.ReactProp;
+import com.swmansion.rnscreens.events.ScreenAppearEvent;
+import com.swmansion.rnscreens.events.ScreenDisappearEvent;
+import com.swmansion.rnscreens.events.ScreenDismissedEvent;
+import com.swmansion.rnscreens.events.ScreenWillAppearEvent;
+import com.swmansion.rnscreens.events.ScreenWillDisappearEvent;
+import com.swmansion.rnscreens.events.StackFinishTransitioningEvent;
+
 import java.util.Map;
 import javax.annotation.Nullable;
